feat : update code und remove unneeded file
