--- frontend/src/components/detail/detail.css
@@ -1,12 +1,0 @@
-.DetailedView {
-    display: flex;
-    flex-direction: column;
-}
-
-.DetailView-image {
-    width: 25rem;
-}
-
-.Detaileview-backButton {
-    width: 5rem;
-}
