+++ frontend/src/components/overall.jsx
@@ -1,15 +1,13 @@
 import React from 'react';
-import { mergeSort } from '../../helpers/sorting.helper';
+import '../assets/styles/index.scss';
 
-
-import './overall.css';
+import { saveInLocalStorage } from './sorting.number';
+import { sorting } from './sorting.number';
 
 const Overall = ({bestOf, setisDetails}) => {
     const paintings = bestOf ? bestOf : [];
 
-    mergeSort(paintings);
-
- 
+    sorting(paintings);
 
     const handleClick = (paintingNumber) => {
         setisDetails(true);
@@ -17,7 +15,7 @@ const Overall = ({bestOf, setisDetails}) => {
     }
 
     const getPaintingsPreview = paintings.map((painting, i) => (
-        <div key={i} className='Preview-painting'>
+        <div key={i} className='Overall-preview'>
             <button onClick={() => handleClick(painting.sortingNumber)}>
                 <img src={painting.images.overall.images[0].sizes.medium.src} height='200px'/>
             </button>
@@ -28,7 +26,7 @@ const Overall = ({bestOf, setisDetails}) => {
     return (
         <div>
             <h1>Dies ist eine Übersicht über die Meisterwerke von Lukas Cranach</h1>
-            <div className='Preview'>
+            <div className='Overall'>
                 {getPaintingsPreview}
             </div>
         </div>
