--- frontend/src/components/overall/overall.css
@@ -1,15 +1,0 @@
-.Preview {
-    display: flex;
-    flex-wrap: wrap;
-}
-
-.Preview-painting {
-    margin: 1.5rem;
-    padding: 0.5rem;
-}
-
-h1 {
-    color:orange ;
-    text-align: center;
-
-}
