--- frontend/src/helpers/localStorage.helper.jsx
@@ -1,7 +1,0 @@
-export const saveInLocalStorage = (key, value) => {
-    localStorage.setItem(key, value);
-};
-
-export const getFromLocalStorage = (key) => {
-    return localStorage.getItem(key);
-}
